[Cond] Move public headers on to interface directory
