+++ CondCore/CondDB/src/Logger.cc
@@ -2,7 +2,7 @@
 #include "CondCore/CondDB/interface/Auth.h"
 #include "CondCore/CondDB/interface/Exception.h"
 //
-#include "DbCore.h"
+#include "../interface/DbCore.h"
 #include "CoralMsgReporter.h"
 #include "RelationalAccess/ITransaction.h"
 //
